[수정]상세페이지 피드백 반영(관심사 분리, 불필요한 코드 삭제 등)
